+++ packages/plugin-defender-hardhat/src/propose-upgrade.ts
@@ -9,16 +9,20 @@ import {
   isTransparentOrUUPSProxy,
 } from '@openzeppelin/upgrades-core';
 import { AdminClient, ProposalResponse } from 'defender-admin-client';
-import type { ContractFactory } from 'ethers';
-import { FormatTypes } from 'ethers/lib/utils';
+import type { ContractFactory, ethers } from 'ethers';
+import { FormatTypes, getContractAddress } from 'ethers/lib/utils';
 import { HardhatRuntimeEnvironment } from 'hardhat/types';
 import { fromChainId } from 'defender-base-client';
+
+export interface ProposalResponseWithUrlAndTx extends ProposalResponse {
+  txResponse?: ethers.providers.TransactionResponse;
+}
 
 export type ProposeUpgradeFunction = (
   proxyAddress: string,
   ImplFactory: ContractFactory,
   opts?: ProposalOptions,
-) => Promise<ProposalResponse>;
+) => Promise<ProposalResponseWithUrlAndTx>;
 
 export interface ProposalOptions extends ValidationOptions {
   title?: string;
@@ -58,9 +62,23 @@ export function makeProposeUpgrade(hre: HardhatRuntimeEnvironment): ProposeUpgra
       await getImplementationAddress(hre.network.provider, proxyAddress);
     }
 
-    const newImplementation = await hre.upgrades.prepareUpgrade(proxyAddress, ImplFactory, moreOpts);
     const contract = { address: proxyAddress, network, abi: ImplFactory.interface.format(FormatTypes.json) as string };
-    return client.proposeUpgrade(
+
+    const prepareUpgradeResult = await hre.upgrades.prepareUpgrade(proxyAddress, ImplFactory, {
+      getTxResponse: true,
+      ...moreOpts,
+    });
+
+    let txResponse, newImplementation;
+
+    if (typeof prepareUpgradeResult === 'string') {
+      newImplementation = prepareUpgradeResult;
+    } else {
+      txResponse = prepareUpgradeResult;
+      newImplementation = getContractAddress(txResponse);
+    }
+
+    const proposalResponse = await client.proposeUpgrade(
       {
         newImplementation,
         title,
@@ -71,5 +89,10 @@ export function makeProposeUpgrade(hre: HardhatRuntimeEnvironment): ProposeUpgra
       },
       contract,
     );
+
+    return {
+      ...proposalResponse,
+      txResponse,
+    };
   };
 }